+++ frontend/src/pages/podcast/PodcastDetailPage/PodcastDetailPage.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useMemo, useState } from "react"
 import { useParams, useSearchParams } from "react-router"
 import { IoReload } from "react-icons/io5"
 import LoadingDisplay from "../../../components/LoadingDisplay/LoadingDisplay.tsx"
-import PodcastEpisodeList from "../../../components/PodcastEpisodeList/PodcastEpisodeList.tsx"
+import PodcastEpisodeList from "../../../features/podcast/episode/PodcastEpisodeList/PodcastEpisodeList.tsx"
 import PodcastCard from "../../../components/PodcastCard/index.tsx"
 import usePodcastEpisodes from "../../../hooks/podcast/usePodcastEpisodes.ts"
 import Pagination from "../../../components/Pagination/Pagination.tsx"
